Basic JWT auth implementation
Verify login and store a signed token on the client within a cookie. Use the cookie to auth subsequent requests.

The JWT secret key is stored in an environment variable. Locally, dotenv-load is used to add these values on to the Lambda's process.env at runtime. On production, these values are added via the AWS console.